+++ src/hdmf/__init__.py
@@ -1,9 +1,9 @@
 from . import query
 from .backends.hdf5.h5_utils import H5Dataset, H5RegionSlicer
-from .container import Container, Data, DataRegion, HERDManager, TermSetWrapper
+from .container import Container, Data, DataRegion, HERDManager
 from .region import ListSlicer
 from .utils import docval, getargs
-from .term_set import TermSet
+from .term_set import TermSet, TermSetWrapper
 
 
 @docval(
