+++ src/hdmf/term_set.py
@@ -3,6 +3,7 @@ import os
 from collections import namedtuple
 from .utils import docval
 import warnings
+import numpy as np
 
 
 class TermSet:
@@ -169,3 +170,76 @@ class TermSet:
         expander.expand_in_place(self.term_schema_path, enum, output_path)
 
         return output_path
+
+class TermSetWrapper:
+    """
+    This class allows any HDF5 group, dataset, or attribute to have a TermSet.
+
+    In HDMF, a group is a Container object, a dataset is a Data object,
+    an attribute can be a reference type to an HDMF object or a base type, e.g., text.
+    """
+    # @docval({'name': 'termset',
+    #          'type': TermSet,
+    #          'doc': 'The TermSet to be used.'},
+    #         {'name': primitive})
+    def __init__(self, **kwargs):
+        self.__item = kwargs['item']
+        self.__termset = kwargs['termset']
+        self.__validate()
+
+    def __validate(self):
+        # check if list, tuple, array, Data
+        from .container import Data # circular import fix
+        if isinstance(self.__item, (list, np.ndarray, tuple, Data)): # TODO: Future ticket on DataIO support
+            values = self.__item
+        # create list if none of those
+        else:
+            values = [self.__item]
+        # iteratively validate
+        bad_values = []
+        for term in values:
+            validation = self.__termset.validate(term=term)
+            if not validation:
+                bad_values.append(term)
+        if len(bad_values)!=0:
+            msg = ('"%s" is not in the term set.' % ', '.join([str(item) for item in bad_values]))
+            raise ValueError(msg)
+
+    @property
+    def item(self):
+        return self.__item
+
+    @property
+    def termset(self):
+        return self.__termset
+
+    @property
+    def dtype(self):
+        return self.__getattr__('dtype')
+
+    def __getattr__(self, val):
+        """
+        This method is to get attributes that are not defined in init.
+        This is when dealing with data and numpy arrays.
+        """
+        if val in ('data', 'shape', 'dtype'):
+            return getattr(self.__item, val)
+
+    def __getitem__(self, val):
+        """
+        This is used when we want to index items.
+        """
+        return self.__item[val]
+
+    def __next__(self):
+        """
+        We want to make sure all iterators are still valid.
+        """
+        return self.__item.__next__()
+
+
+    def __iter__(self):
+        """
+        We want to make sure our wrapped items are still iterable.
+        """
+        return self.__item.__iter__()
